pyproject.toml
[project]
name = "ws.thyrida"
description = "Mailbox configuration UI"
version = "2.0.0"

dependencies = [
    'pyramid',
    'pyramid_jinja2',
    'pyramid_tm',
    'sqlalchemy >= 2.0.0',
    'transaction',
    'wtforms',
    'zope.component',
    'zope.interface',
    'zope.sqlalchemy',
]
optional-dependencies = {test=[
  "pytest",
  "webtest",
]}

authors = [{name="Wolfgang Schnerring", email="wosc@wosc.de"}]
license = {text=" BSD-3-Clause"}
urls = {Repository="https://github.com/wosc/thyrida"}
requires-python = ">=3.7"

[project.entry-points."paste.app_factory"]
main = "ws.thyrida.application:app_factory"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ws"]
